Organize fragments into a 'fragments' folder
