+++ app/src/main/java/com/github/mateo762/myapplication/MainActivity.kt
@@ -8,6 +8,10 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.core.view.GravityCompat
 import androidx.drawerlayout.widget.DrawerLayout
+import com.github.mateo762.myapplication.fragments.CalendarFragment
+import com.github.mateo762.myapplication.fragments.PicturesFragment
+import com.github.mateo762.myapplication.fragments.ProfileFragment
+import com.github.mateo762.myapplication.fragments.SettingsFragment
 import com.google.android.material.navigation.NavigationView
 import com.google.android.material.navigation.NavigationView.OnNavigationItemSelectedListener
