+++ app/src/main/java/com/github/mateo762/myapplication/fragments/SettingsFragment.kt
@@ -1,10 +1,11 @@
-package com.github.mateo762.myapplication
+package com.github.mateo762.myapplication.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.github.mateo762.myapplication.R
 
 class SettingsFragment : Fragment() {
